server/workers/tts/requirements.txt
piper-tts==1.2.0
redis==5.2.1
flask==3.1.1
